Enhance parsing map and finish norme

+++ sources/raycaster/raycast_cal.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   raycast_cal.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: antgabri <antgabri@student.42.fr>          +#+  +:+       +#+        */
+/*   By: monsieurc <monsieurc@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/25 14:50:15 by monsieurc         #+#    #+#             */
-/*   Updated: 2024/05/29 11:28:30 by antgabri         ###   ########.fr       */
+/*   Updated: 2024/05/29 14:19:56 by monsieurc        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ t_vector2	get_coord_texture(t_ray *ray, t_data *data)
 		|| (ray->hit_axis == 1 && ray->dir.y < 0))
 		texture.x = data->texture[0]->size.x - texture.x;
 	ray->step_texture_wall = 1.0 * data->texture[0]->size.y / ray->line_height;
-	texture.y = (ray->draw_start -(WIN_HEIGHT / 2)
+	texture.y = (int)(ray->draw_start -(WIN_HEIGHT / 2)
 			+ (ray->line_height / 2)) * ray->step_texture_wall;
 	return (texture);
 }
